Make BN states public
