+++ chainerx_cc/chainerx/cuda/cuda_device/batch_norm.cc
@@ -12,6 +12,7 @@
 #include "chainerx/array.h"
 #include "chainerx/axes.h"
 #include "chainerx/backend_util.h"
+#include "chainerx/cuda/cuda_device.h"
 #include "chainerx/cuda/cuda_set_device_scope.h"
 #include "chainerx/cuda/cudnn.h"
 #include "chainerx/cuda/op_regist.h"
@@ -69,23 +70,6 @@ cuda_internal::CudnnTensorDescriptor DeriveBatchNormTensorDescriptor(
     CheckCudnnError(cudnnDeriveBNTensorDescriptor(*derive_desc, *x_desc, mode));
     return derive_desc;
 }
-
-struct CudaBatchNormGradState : public BatchNormGradState {
-public:
-    CudaBatchNormGradState(Array x_cont, Array x_mean, Array x_inv_std, Dtype beta_dtype)
-        : x_cont_{std::move(x_cont)}, x_mean_{std::move(x_mean)}, x_inv_std_{std::move(x_inv_std)}, beta_dtype_{beta_dtype} {}
-
-    const Array& x_cont() const { return x_cont_; }
-    const Array& x_mean() const { return x_mean_; }
-    const Array& x_inv_std() const { return x_inv_std_; }
-    Dtype beta_dtype() const { return beta_dtype_; }
-
-private:
-    Array x_cont_;
-    Array x_mean_;
-    Array x_inv_std_;
-    Dtype beta_dtype_;
-};
 
 class CudaBatchNormOp : public BatchNormOp {
 public:
